Commiting

+++ main.h
@@ -3,8 +3,9 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 int is_prime(int nb);
-int find_pair(long int nb);
+int find_pair(unsigned long long int nb);
 
 #endif /* RSA */
